+++ provider/ansible/property_override.rb
@@ -13,7 +13,7 @@
 
 require 'api/object'
 require 'provider/abstract_core'
-require 'provider/property_override'
+require 'provider/overrides/resources'
 
 module Provider
   module Ansible
@@ -25,7 +25,7 @@ module Provider
     end
 
     # Ansible-specific overrides to api.yaml.
-    class PropertyOverride < Provider::PropertyOverride
+    class PropertyOverride < Provider::Overrides::PropertyOverride
       include OverrideFields
       def validate
         super
