+++ provider/inspec/resource_override.rb
@@ -12,7 +12,7 @@
 # limitations under the License.
 
 require 'provider/core'
-require 'provider/resource_override'
+require 'provider/overrides/resources'
 
 module Provider
   class Inspec < Provider::Core
@@ -29,7 +29,7 @@ module Provider
     end
 
     # Product specific overriden properties for inspec
-    class ResourceOverride < Provider::ResourceOverride
+    class ResourceOverride < Provider::Overrides::ResourceOverride
       include OverrideProperties
 
       def validate
