+++ provider/terraform/custom_code.rb
@@ -16,7 +16,7 @@ require 'api/object'
 require 'compile/core'
 require 'google/golang_utils'
 require 'provider/abstract_core'
-require 'provider/property_override'
+require 'provider/overrides/resources'
 
 module Provider
   class Terraform < Provider::AbstractCore
