+++ provider/terraform/resource_override.rb
@@ -12,7 +12,7 @@
 # limitations under the License.
 
 require 'provider/abstract_core'
-require 'provider/resource_override'
+require 'provider/overrides/resources'
 require 'provider/terraform/custom_code'
 
 module Provider
@@ -42,7 +42,7 @@ module Provider
 
     # A class to control overridden properties on terraform.yaml in lieu of
     # values from api.yaml.
-    class ResourceOverride < Provider::ResourceOverride
+    class ResourceOverride < Provider::Overrides::ResourceOverrides
       include OverrideProperties
 
       def validate
